FeatureComplete/Check: split parsing of configuration off to separate class
Changes in the functionality now contained in the `Config` class:
* All properties are now `private`.
* A magic `__get()` and `__isset()` method is available to access the value of the properties.
* A magic `__set()` and `__unset()` method have been added to _prevent_ any changes to the value of the properties from outside the class and prevent dynamic properties from being set.
* The `showVersion()` method is now `public` as it needs to be called from the `Check` class.
* The `showHelp()` method is now `private`.

Includes:
* Re-ordering the properties.
